Restructure to use lib.rs
